KERML_-117 Moved adding index result type to new IndexExpressionAdapter.
